requirements.txt
astroid==1.4.8
coverage==4.2
Django==1.10.3
django-guardian==1.4.6
django-model-utils==2.6
djangorestframework==3.5.3
flake8==3.2.1
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.5.2
mysqlclient==1.3.9
pycodestyle==2.2.0
pyflakes==1.3.0
pylint==1.6.4
PyYAML==3.12
requests==2.12.3
six==1.10.0
wrapt==1.10.8
